requirements.txt
streamlit
pandas
numpy
matplotlib.pyplot
altair